Benchmarks Website V3: Admin and Auto-Deploy (#7849)
## Summary

http://ec2-18-219-54-101.us-east-2.compute.amazonaws.com:3000/

Hopefully the Benchmarks V3 finale. The website is one Rust binary + one
DuckDB file managed by
systemd: `vortex-bench-deploy.timer` polls `origin/develop` every 60s,
rebuilds + atomic-swaps the
binary on website-touching commits, `/health`-verifies, rolls back on
failure.
`vortex-bench-backup.timer` fires hourly, calls a new loopback-only
`POST /api/admin/snapshot`
(per-table Vortex export under one `READ ONLY` transaction), `tar czf`s
the result, and uploads to
`s3://vortex-benchmark-results-database/v3-backups/<ts>.tar.gz`. Admin
endpoints (snapshot +
read-only `POST /api/admin/sql`) sit on a separate
`ADMIN_BEARER_TOKEN`-gated listener that refuses
to start on a non-loopback bind. v2 keeps shipping unchanged via the
existing publish workflow until
DNS flips.

Operator runbook (one-time install on a fresh EC2 host, day-to-day,
recovery):
[`benchmarks-website/ops/README.md`](benchmarks-website/ops/README.md).
On a clean box, `git clone`
the repo, edit `INGEST_BEARER_TOKEN` + `ADMIN_BEARER_TOKEN` in
`/etc/vortex-bench.env`,
`./benchmarks-website/ops/install.sh`.

<details>
<summary>Admin cheat sheet (generated by claude)</summary>

### Status — "what's the state right now?"

```bash
systemctl status vortex-bench-server                    # main service status + recent log lines
systemctl list-timers vortex-bench-deploy.timer vortex-bench-backup.timer
                                                         # next + last fire of each timer
curl -fsS http://127.0.0.1:3000/health | jq             # live build_sha, row counts, db path
cat /var/lib/vortex-bench/last-deployed-sha             # SHA the deploy timer thinks is live
readlink /var/lib/vortex-bench/bin/vortex-bench-server  # symlink target = versioned binary
```

### Manual triggers — "do something now"

```bash
/var/lib/vortex-bench/ops/restart.sh                    # restart binary in place, show before/after
/var/lib/vortex-bench/ops/force-rebuild.sh              # rebuild origin/$DEPLOY_BRANCH + restart
sudo systemctl start vortex-bench-deploy.service        # run deploy IF origin moved (no-op otherwise)
sudo systemctl start vortex-bench-backup.service        # snapshot to S3 right now
/var/lib/vortex-bench/ops/migrate.sh run --output \
    /var/lib/vortex-bench/bench.duckdb                  # re-run v2→v3 migration
/var/lib/vortex-bench/ops/inspect.sh "SELECT COUNT(*) FROM commits;"  # read-only SQL
```

### Logs — "what happened?"

```bash
journalctl -fu vortex-bench-server                      # live server logs
journalctl -fu vortex-bench-deploy.service              # watch a deploy in progress
journalctl -u vortex-bench-deploy.service --since '15 min ago' --no-pager
journalctl -u vortex-bench-backup.service --since '4 hours ago' --no-pager
journalctl -u vortex-bench-server --since '5 min ago' --no-pager
```

### Pause / resume the autopilot — "stop fiddling for a minute"

```bash
sudo systemctl stop vortex-bench-deploy.timer           # pause auto-deploys
sudo systemctl start vortex-bench-deploy.timer          # resume

sudo systemctl stop vortex-bench-backup.timer           # pause hourly snapshots
sudo systemctl start vortex-bench-backup.timer          # resume

sudo systemctl stop vortex-bench-server                 # take the site down
sudo systemctl start vortex-bench-server                # bring it back
```

### Config changes — "edit settings"

```bash
sudo $EDITOR /etc/vortex-bench.env                      # tokens, DEPLOY_BRANCH, S3 prefix, etc.
sudo systemctl restart vortex-bench-server              # pick up server-side env changes
                                                         # (no rebuild — restart.sh shows new pid)
/var/lib/vortex-bench/ops/force-rebuild.sh              # pick up build-time env changes
```

### Recover from S3 — "restore a backup"

```bash
aws s3 ls s3://vortex-benchmark-results-database/v3-backups/ | tail
LATEST=$(aws s3 ls s3://vortex-benchmark-results-database/v3-backups/ | tail -1 | awk '{print $4}')
aws s3 cp "s3://vortex-benchmark-results-database/v3-backups/$LATEST" /tmp/
cd /tmp && tar xzf "$LATEST"
ts="${LATEST%.tar.gz}"
sudo systemctl stop vortex-bench-server
sudo -u ec2-user rm -f /var/lib/vortex-bench/bench.duckdb \
                       /var/lib/vortex-bench/bench.duckdb.wal
duckdb /var/lib/vortex-bench/bench.duckdb <<EOF
INSTALL vortex;
LOAD vortex;
.read /tmp/${ts}/schema.sql
INSERT INTO commits             SELECT * FROM read_vortex('/tmp/${ts}/commits.vortex');
INSERT INTO query_measurements  SELECT * FROM read_vortex('/tmp/${ts}/query_measurements.vortex');
INSERT INTO compression_times   SELECT * FROM read_vortex('/tmp/${ts}/compression_times.vortex');
INSERT INTO compression_sizes   SELECT * FROM read_vortex('/tmp/${ts}/compression_sizes.vortex');
INSERT INTO random_access_times SELECT * FROM read_vortex('/tmp/${ts}/random_access_times.vortex');
INSERT INTO vector_search_runs  SELECT * FROM read_vortex('/tmp/${ts}/vector_search_runs.vortex');
EOF
sudo systemctl start vortex-bench-server
```

The 5 commands you'll actually use day-to-day: `restart.sh`,
`force-rebuild.sh`, `inspect.sh`,
`journalctl -fu …`, `curl /health | jq`. Everything else is rare.

</details>

---------

+++ vortex-bench/src/v3.rs
@@ -349,7 +349,11 @@ pub fn query_measurement_record(
         dataset,
         dataset_variant,
         scale_factor,
-        query_idx: u32::try_from(qm.query_idx).unwrap_or(u32::MAX),
+        // Clamp at `i32::MAX`, not `u32::MAX`: the server's `query_idx`
+        // field is `i32`, so a saturated `u32::MAX` would fail serde
+        // deserialization. Real `query_idx` values are 0..200, so the
+        // saturation path is defensive rather than load-bearing.
+        query_idx: u32::try_from(qm.query_idx).unwrap_or(i32::MAX as u32),
         storage: qm.storage.clone(),
         engine: engine_label(qm.target.engine).to_string(),
         format: qm.target.format.name().to_string(),
@@ -447,7 +451,10 @@ pub fn vector_search_record(
     rows_scanned: u64,
     bytes_scanned: u64,
 ) -> V3Record {
-    let iterations = u32::try_from(all_runs_ns.len()).unwrap_or(u32::MAX);
+    // Clamp at `i32::MAX`: server-side `iterations` is `i32` (see the
+    // same-named field on `vortex_bench_server::records::VectorSearchRun`),
+    // so saturating to `u32::MAX` would 400 the envelope.
+    let iterations = u32::try_from(all_runs_ns.len()).unwrap_or(i32::MAX as u32);
     V3Record::VectorSearchRun(VectorSearchRunRecord {
         commit_sha: GIT_COMMIT_ID.clone(),
         dataset: dims.dataset.to_string(),
@@ -485,8 +492,16 @@ pub fn write_jsonl_to_path(path: &std::path::Path, records: &[V3Record]) -> std:
     write_jsonl(&mut file, records)
 }
 
+/// Convert a `Duration` to `u64` nanoseconds, clamping at the largest
+/// value the server-side `value_ns: i64` deserializer can accept.
+///
+/// The wire field is `u64` (see `value_ns` on every record type below),
+/// but the server's [`vortex_bench_server::records`] mirrors them as `i64`.
+/// Without the clamp, an overflowed measurement would land at `u64::MAX`
+/// here, fail serde deserialization on the server, and 400 the whole
+/// ingest envelope.
 fn duration_as_ns(d: std::time::Duration) -> u64 {
-    u64::try_from(d.as_nanos()).unwrap_or(u64::MAX)
+    u64::try_from(d.as_nanos()).unwrap_or(i64::MAX as u64)
 }
 
 fn engine_label(engine: Engine) -> &'static str {
